Combine airship bindings
