+++ src/AirshipBindings.iOS/AirshipBindings.iOS.csproj
@@ -57,5 +57,9 @@
       <Kind>Framework</Kind>
       <SmartLink>False</SmartLink>
     </NativeReference>
+    <NativeReference Include="..\..\Carthage\Build\iOS\AirshipAppExtensions.framework">
+      <Kind>Framework</Kind>
+      <SmartLink>False</SmartLink>
+    </NativeReference>
   </ItemGroup>
 </Project>
