+++ src/AirshipBindings.iOS/ApiDefinition.cs
@@ -9,6 +9,60 @@ using ObjCRuntime;
 using UIKit;
 using UserNotifications;
 
+namespace AirshipAppExtensions
+{
+	// @interface UAMediaAttachmentExtension : UNNotificationServiceExtension
+	[BaseType(typeof(UNNotificationServiceExtension))]
+	[Preserve]
+	interface UAMediaAttachmentExtension
+	{
+		// -(NSString *)uniformTypeIdentifierForData:(NSData *)data;
+		[Export("uniformTypeIdentifierForData:")]
+		string UniformTypeIdentifier(NSData data);
+	}
+
+	// @interface UAMediaAttachmentContent : NSObject
+	[BaseType(typeof(NSObject))]
+	[Preserve]
+	interface UAMediaAttachmentContent
+	{
+		// @property (readonly, nonatomic) NSString * body;
+		[Export("body")]
+		string Body { get; }
+
+		// @property (readonly, nonatomic) NSString * title;
+		[Export("title")]
+		string Title { get; }
+
+		// @property (readonly, nonatomic) NSString * subtitle;
+		[Export("subtitle")]
+		string Subtitle { get; }
+	}
+
+	// @interface UAMediaAttachmentPayload : NSObject
+	[BaseType(typeof(NSObject))]
+	[Preserve]
+	interface UAMediaAttachmentPayload
+	{
+		// +(instancetype)payloadWithJSONObject:(id)object;
+		[Static]
+		[Export("payloadWithJSONObject:")]
+		UAMediaAttachmentPayload Payload(NSObject @object);
+
+		// @property (readonly, nonatomic) NSMutableArray * urls;
+		[Export("urls")]
+		NSMutableArray Urls { get; }
+
+		// @property (readonly, nonatomic) NSDictionary * options;
+		[Export("options")]
+		NSDictionary Options { get; }
+
+		// @property (readonly, nonatomic) UAMediaAttachmentContent * content;
+		[Export("content")]
+		UAMediaAttachmentContent Content { get; }
+	}
+}
+
 namespace UrbanAirship {
 
 	// typedef void (^UAJavaScriptDelegateCompletionHandler)(NSString * _Nullable);
