--- src/AirshipExtensionBindings.iOS/AirshipExtensionBindings.iOS.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AirshipExtensionBindings.iOS", "AirshipExtensionBindings.iOS\AirshipExtensionBindings.iOS.csproj", "{2C453988-C24E-431A-A17C-6989E93C795E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2C453988-C24E-431A-A17C-6989E93C795E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2C453988-C24E-431A-A17C-6989E93C795E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2C453988-C24E-431A-A17C-6989E93C795E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2C453988-C24E-431A-A17C-6989E93C795E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
