--- src/AirshipExtensionBindings.iOS/AirshipExtensionBindings.iOS/Properties/AssemblyInfo.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-
-using Foundation;
-
-// This attribute allows you to mark your assemblies as “safe to link”. 
-// When the attribute is present, the linker—if enabled—will process the assembly 
-// even if you’re using the “Link SDK assemblies only” option, which is the default for device builds.
-
-[assembly: LinkerSafe]
-
-[assembly: AssemblyTitle("AirshipExtensionBindings")]
-
-// Information about this assembly is defined by the following attributes. 
-// Change them to the values specific to your project.
