--- src/AirshipExtensionBindings.iOS/AirshipExtensionBindings.iOS/Structs.cs
@@ -1,5 +1,0 @@
-﻿using System;
-
-namespace AirshipAppExtensions
-{
-}
